Fix - Parameters Type

+++ kepleruniversal.h
@@ -26,7 +26,7 @@
  * @note keplerUniversal function - original file keplerUniversal.m
  */
 //------------------------------------------------------------------------------
-void KeplerUniversal(double m, double **r0, double **v0, double *t, double mu, double **r, double **v, double *mm);
+void KeplerUniversal(int m, double **r0, double **v0, double *t, double mu, double **r, double **v, int *mm);
 
 //------------------------------------------------------------------------------
 // function C2c3
@@ -41,6 +41,6 @@ void KeplerUniversal(double m, double **r0, double **v0, double *t, double mu, d
  * @note c2c3 function - original file keplerUniversal.m
  */
 //------------------------------------------------------------------------------
-void C2c3(double m, double psi, double *c2, double *c3, double *mm);
+void C2c3(int m, double psi, double *c2, double *c3, int *mm);
 
 #endif //CODIGO_KEPLERUNIVERSAL_H
